+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Wallet position on screen is now remembered for the next time it's opened
 - Added a warning on balance reminding it might not be up-to-date when the wallet is not fully synchronized
 - Added ability to disable the Transaction Chart.
+- Added Portuguese Language.
 
 #### Adjustments